Safari-609.1.11.1/LayoutTests/fullscreen/requestFullscreen-escape-key-expected.txt

CONSOLE MESSAGE: line 11: The Escape key may not be used as a user gesture to enter fullscreen
Tests that pressing the Escape key cannot be used as user interaction to enter fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.webkitFullscreenEnabled is true
Pressing Escape key....
keydown - Requesting fullscreen
PASS document.webkitFullscreenElement is null
PASS successfullyParsed is true

TEST COMPLETE
